+++ R/Koboconnect.R
@@ -19,7 +19,7 @@
 #'
 #'
 #'
-#' @importFrom httr GET add_headers content progress
+#' @importFrom httr GET add_headers content progress stop_for_status warn_for_status
 #' @importFrom jsonlite fromJSON
 #' @import R6
 #' @import curl
@@ -45,8 +45,7 @@ kobotools_api<- function(url="kobo.humanitarianresponse.info", simplified=TRUE, 
   fullurl<-paste0("https://",url,"/api/v2/assets.json")
   respon.api<-GET(fullurl, authenticate(uname, pwd), progress())
 
-  if(respon.api$status_code!=200) stop(paste0("Error in GET response. Expected 200, recieved ",respon.api$status_code))
-
+  stop_for_status(respon.api, "extract asset details")
   parsed <- fromJSON(content(respon.api, "text", encoding = encoding), simplifyVector = FALSE)
 
   if(simplified==FALSE){
@@ -95,7 +94,7 @@ kobotools_api<- function(url="kobo.humanitarianresponse.info", simplified=TRUE, 
 #' @param pwd is the password of the account
 #' @param encoding is the encoding to be used. Default is "UTF-8".
 #'
-#' @return The function returns the data
+#' @return The function returns the data in json format
 #'
 #'
 #' @importFrom httr GET content authenticate progress
@@ -119,7 +118,7 @@ kobotools_kpi_data<- function(assetid,url="kobo.humanitarianresponse.info", unam
 
   fullurl<-paste0("https://",url,"/api/v2/assets/",assetid,"/data/")
   respon.kpi<-GET(fullurl, authenticate(uname, pwd), progress())
-  if(respon.kpi$status_code!=200) stop(paste0("Error in GET response. Expected 200, recieved ",respon.kpi$status_code))
+  stop_for_status(respon.kpi, "extract data")
   dt<-content(respon.kpi, encoding = encoding)
   return(dt)
 }
@@ -143,6 +142,7 @@ kobotools_kpi_data<- function(assetid,url="kobo.humanitarianresponse.info", unam
 #' @importFrom httr GET content authenticate progress
 #' @importFrom jsonlite fromJSON
 #'
+#'
 #' @export
 
 get_kobo_token <- function(url="kobo.humanitarianresponse.info", uname="", pwd="", encoding = "UTF-8"){
@@ -156,7 +156,7 @@ get_kobo_token <- function(url="kobo.humanitarianresponse.info", uname="", pwd="
 
   fullurl<-paste0("https://",url,"/token/?format=json")
   respon.token<-GET(fullurl, authenticate(uname, pwd), progress())
-  if(respon.token$status_code!=200) stop(paste0("Error in GET response. Expected 200, recieved ",respon.token$status_code))
+  stop_for_status(respon.token,"extract token")
   tkn<-fromJSON(content(respon.token,"text", encoding = encoding))
   return(tkn)
 }
@@ -175,7 +175,7 @@ get_kobo_token <- function(url="kobo.humanitarianresponse.info", uname="", pwd="
 #' @param pwd is the password of the account
 #' @param encoding is the encoding to be used. Default is "UTF-8".
 #'
-#' @return The function returns the token associated with your id and password in the given url.
+#' @return The function returns a list of exports available for the account id and password entered.
 #'
 #'
 #' @importFrom httr GET content authenticate progress
@@ -194,7 +194,7 @@ kobo_exports <- function(url="kobo.humanitarianresponse.info", uname="", pwd="",
 
   fullurl<-paste0("https://",url,"/exports/")
   respon.exp<-GET(fullurl, authenticate(uname, pwd), progress())
-  if(respon.exp$status_code!=200) stop(paste0("Error in GET response. Expected 200, recieved ",respon.exp$status_code))
+  stop_for_status(respon.exp,"extract export list.")
   exports<-fromJSON(content(respon.exp,"text", encoding = encoding))
   return(exports)
 }
@@ -212,17 +212,30 @@ kobo_exports <- function(url="kobo.humanitarianresponse.info", uname="", pwd="",
 #' @param uname is username of your kobotoolbox account
 #' @param pwd is the password of the account
 #' @param assetid is the id of the asset for which the export is to be created
-#' @param type is the type of export to be created. For e.g. 'xls' or 'csv'.
-#' @param all takes logical value in string format. "true" indicates the field from all versions of the asset to be used. Defaults to "false".
+#' @param all takes logical value in string format. Used to specify whether fields from all form versions will be included in the export.Acceptable values are "true" or "false". Default value is "false".
 #' @param lang takes the language. For e.g. "English (en)".
-#' @param hierarchy takes logical value in string format. "true" indicates hierarchy will be used in all labels. Default value is "false".
+#' @param hierarchy takes logical value in string format. Used to specify whether the group hierarchy will be displayed in labels. Acceptable values are "true" or "false". Default value is "false".
 #' @param grp_sep is the group separator. Default value is "/".
 #' @param include_grp defines whether or not to include groups. Default value is "true".
+#' @param multi_sel is used to specify the display of multiple_select-type responses. Valid inputs include "both", "summary" or "details". Default is "both".
+#' @param fields is an array of column names to be included in the export (including their group hierarchy). Valid inputs include:
+#' An array containing any string value that matches the XML column name,
+#' An empty array which will result in all columns being included,
+#' If "fields" is not included in the "export_settings", all columns will be included in the export
 #'
-#' @return The function returns the token associated with your id and password in the given url.
+#' @param media_url This will include an additional column for media-type questions ("question_name_URL") with the URL link to the hosted file. Valid inputs are "true" or "false". Default value is true.
+#' @param sub_ids is an array of submission ids that will filter exported submissions to only the specified array of ids. Valid inputs include an array containing integer values or an empty array.#'
+#' @param type specifies the export format. Valid formats include "csv","xls", "spss_labels" and "geojson"
+#' @param qry is a JSON object containing a Mongo filter query for filtering exported submissions. Valid inputs include
+#' a JSON object containing a valid Mongo query or An empty JSON object (no filtering)
+#' @param flatten is a is a boolean (in form of string) value and only relevant when exporting to "geojson" format. Valid inputs are "true" and "false"
 #'
 #'
-#' @importFrom httr POST content authenticate progress
+#'
+#' @return The function creates an export, prints and returns the URL of the export created
+#'
+#'
+#' @importFrom httr POST content authenticate progress GET
 #' @importFrom jsonlite fromJSON
 #'
 #' @export
@@ -230,71 +243,80 @@ kobo_exports <- function(url="kobo.humanitarianresponse.info", uname="", pwd="",
 
 kobo_export_create <- function(url="kobo.humanitarianresponse.info", uname="", pwd="",
                                assetid="", type= "csv", all="false", lang="_default",
-                               hierarchy="false", include_grp="true",grp_sep="/"){
+                               hierarchy="false", include_grp="true",grp_sep="/",
+                               multi_sel="both", fields=NULL, media_url ="true",
+                               sub_ids=NULL, qry=NULL, flatten="true"){
 
-  if(!is.character(url)) stop("URL entered is not a string")
-  if(!is.character(uname)) stop("uname (username) entered is not a string")
-  if(!is.character(pwd)) stop("pwd (password) entered is not a string")
-  if(!is.character(assetid)) stop("assetid (asset id) entered is not a string")
-  if(!is.character(type)) stop("type entered is not a string")
-  if(!is.character(all)) stop("all entered is not a string")
-  if(!is.character(lang)) stop("lang entered is not a string")
-  if(!is.character(hierarchy)) stop("hierarchy entered is not a string")
-  if(!is.character(grp_sep)) stop("grp_sep entered is not a string")
-  if(!is.character(include_grp)) stop("include_grp entered is not a string")
+  export_res<-export_creator(url=url, uname=uname, pwd=pwd,
+                 assetid=assetid, type= type, all=all, lang=lang,
+                 hierarchy=hierarchy, include_grp=include_grp,grp_sep=grp_sep,
+                 multi_sel=multi_sel, fields=fields, media_url=media_url,
+                 sub_ids=sub_ids, qry=qry, flatten=flatten)
 
-  if(is.null(url)) stop("URL empty")
-  if(is.null(uname)) stop("uname (username) empty")
-  if(is.null(pwd)) stop("pwd (password) empty")
-  if(is.null(assetid)) stop("assetid (asset id) empty")
-  if(is.null(type)) stop("type empty")
-  if(is.null(all)) stop("all empty")
-  if(is.null(lang)) stop("lang empty")
-  if(is.null(hierarchy)) stop("hierarchy empty")
-  if(is.null(grp_sep)) stop("grp_sep empty")
-  if(is.null(include_grp)) stop("include_grp empty")
-
-  pre_export<- kobo_exports(url=url, uname=uname, pwd=pwd)
-  pre_count<-as.integer(pre_export$count)
-
-  fullurl<-paste0("https://",url,"/exports/")
-  task<-POST(fullurl, authenticate(uname, pwd),
-             body=list(
-               source=paste0("https://",url,"/assets/",assetid,"/"),
-               type=type, fields_from_all_versions= all, lang=lang, hierarchy_in_labels = hierarchy,
-               include_groups=include_grp,
-               group_sep=grp_sep
-             ),
-             progress())
-  if(task$status_code!=201) stop(paste0("Error in GET response. Expected 201, recieved ",task$status_code))
-
-  if(task$status_code==201) cat("Export instruction sent succesfully. Waiting for result. \n")
-
-  post_export<-kobo_exports(url=url, uname=uname, pwd=pwd)
-  post_count<-as.integer(post_export$count)
-
-  while(post_count<=pre_count){
-    cat("\n ...Execution in Progress \n")
-    post_export<-kobo_exports(url=url, uname=uname, pwd=pwd)
-    post_count<-as.integer(post_export$count)
-  }
-  rm(pre_export)
-  rm(pre_count)
-
-  if(post_export$results$status[post_count]=="error"){
-    stop(paste0("Did not execute. Encountered ",post_export$results$messages$error_type[post_count],". ",
-           post_export$results$messages$error[post_count],". \n"))
-  }
-
-  print("Export successful")
-  # print(post_count)
-  while(is.na(post_export$results$result[post_count])){
-    # print("waiting.. \n")
-    post_export<-kobo_exports(url=url, uname=uname, pwd=pwd)
-  }
-  print(post_export$results$result[post_count])
-  return(post_export$results$result[post_count])
-
+  print(export_res[1])
+  return(export_res[1])
 }
 
+
+#' Creates a data frame after creating a 'csv' export and downloading it
+#'
+#'@description
+#' `kobo_df_download` is a wrapper for kobotoolbox API `https://[url]/exports/..`
+#'
+#'@details
+#' The function creates an export of survey data in 'csv'. If successful, it attempts to download the data and and return a data frame.
+#'
+#' @param url The `[url]` of kobotoolbox Default is "kobo.humanitarianresponse.info".
+#' @param uname is username of your kobotoolbox account
+#' @param pwd is the password of the account
+#' @param assetid is the id of the asset for which the export is to be created
+#' @param all takes logical value in string format. Used to specify whether fields from all form versions will be included in the export.Acceptable values are "true" or "false". Default value is "false".
+#' @param lang takes the language. For e.g. "English (en)".
+#' @param hierarchy takes logical value in string format. Used to specify whether the group hierarchy will be displayed in labels. Acceptable values are "true" or "false". Default value is "false".
+#' @param grp_sep is the group separator. Default value is "/".
+#' @param include_grp defines whether or not to include groups. Default value is "true".
+#' @param fsep is the separator of the downloaded csv file. In most of the cases, it is ";", which is the default. However,
+#' if you notice that the separator is the downloaded data is "," or something else, you can change it.
+#' @param multi_sel is used to specify the display of multiple_select-type responses. Valid inputs include "both", "summary" or "details". Default is "both".
+#' @param fields is an array of column names to be included in the export (including their group hierarchy). Valid inputs include:
+#' An array containing any string value that matches the XML column name,
+#' An empty array which will result in all columns being included,
+#' If "fields" is not included in the "export_settings", all columns will be included in the export
+#'
+#' @param media_url This will include an additional column for media-type questions ("question_name_URL") with the URL link to the hosted file. Valid inputs are "true" or "false". Default value is true.
+#' @param sub_ids is an array of submission ids that will filter exported submissions to only the specified array of ids. Valid inputs include an array containing integer values or an empty array.
+#' @param sleep is the sleep time between API actions. For example, it takes time to download an export. But R does not wait for the download to finish before going to next step. Hence the need to provide a break between consecutive API actions. Default value is 2 (seconds).
+#'
+#' @return The function returns a data frame of data downloaded from 'Kobotoolbox'.
+#'
+#'
+#' @importFrom httr POST content authenticate progress DELETE GET
+#' @importFrom jsonlite fromJSON
+#' @importFrom utils read.csv
+#'
+#' @export
+
+
+kobo_df_download <- function(url="kobo.humanitarianresponse.info", uname="", pwd="",
+                               assetid="", all="false", lang="_default",
+                               hierarchy="false", include_grp="true",grp_sep="/",fsep=";",
+                              multi_sel="both", media_url="true", fields=NULL, sub_ids=NULL, sleep=2){
+
+  new_export_details<-export_creator(url=url, uname=uname, pwd=pwd,
+                                                 assetid=assetid, type= "csv", all=all, lang=lang,
+                                                 hierarchy=hierarchy, include_grp=include_grp,grp_sep=grp_sep,
+                                     multi_sel=multi_sel, fields=fields, media_url=media_url, sub_ids=sub_ids)
+
+  Sys.sleep(sleep)
+  dff<-export_downloader(new_export_details[[1]],fsep, uname, pwd, sleep)
+
+  deleteact<-DELETE(url=paste0(url,"/api/v2/assets/",assetid,"/exports/",new_export_details[[2]],"/"),
+                    authenticate(user=uname, password =pwd), progress())
+  while(is.na(deleteact$status_code) | is.null(deleteact$status_code)){
+    print("Attempting export deletion \n")
+  }
+  warn_for_status(deleteact,"delete export. Please delete manually.")
+  if(deleteact$status_code==204) print("Export deleted from server")
+  return(dff)
+}