+++ NEWS.md
@@ -4,4 +4,5 @@
 * Eliminated old functions that used older version of 'Kobotoolbox' APIs
 * Simplified authentication. Instead of using API token, now only user id and password is sufficient
 * Added a `NEWS.md` file to track changes to the package.
+* Supports export creation to enable download of 'csv' and 'xls' files.